% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getDf}
\alias{getDf}
\title{Get the data frame from the network}
\usage{
getDf(net)
}
\arguments{
\item{net}{network obtained from ARACNe}
}
\value{
data frame
}
\description{
This function is used to get a data frame from the file obtained by the
network reconstruction with ARACNe.
}
\examples{

# Load the example network
data(network)

# Get the data frame from the network obtained by ARACNe
net_df <- getDf(network)

}
