% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getGeneSymbolDataframe}
\alias{getGeneSymbolDataframe}
\title{Convert gene Ensembl IDs in the corresponding gene symbols}
\usage{
getGeneSymbolDataframe(dataf, symbols, organism = "")
}
\arguments{
\item{dataf}{data frame obtained by the network}

\item{symbols}{data frame with the gene IDs and the corresponding symbols}

\item{organism}{hsapiens or mmusculus}
}
\value{
data frame
}
\description{
This function is used to convert the gene IDs of the data frame in the
corresponding gene symbols.
}
\examples{

# Load the example data frame
data(network_df)

# Get the gene symbols corresponding to the gene IDs in the data frame
gene_symbols <- getGeneSymbols(network_df, organism = "hsapiens")

# Get the converted data frame
converted_df <- getGeneSymbolDataframe(network_df, gene_symbols, 
organism = "hsapiens")

}
