% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getGeneSymbolMatrix}
\alias{getGeneSymbolMatrix}
\title{Convert gene Ensembl IDs in the corresponding gene symbols}
\usage{
getGeneSymbolMatrix(vsd, symbols, organism = "")
}
\arguments{
\item{vsd}{matrix containing the normalized expression values}

\item{symbols}{data frame with the gene IDs and the corresponding symbols}

\item{organism}{hsapiens or mmusculus}
}
\value{
matrix to be used with VIPER
}
\description{
This function is used to convert the gene IDs of the expression matrix in
the corresponding gene symbols.
}
\examples{

# Load the example data frame
data(network_df)

# Load the example matrix
data(vsd_matrix)

# Get the gene symbols corresponding to the gene IDs in the data frame
gene_symbols <- getGeneSymbols(network_df, organism = "hsapiens")

# Get the converted matrix
converted_matrix <- getGeneSymbolMatrix(vsd_matrix, gene_symbols,
organism = "hsapiens")

}
