% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getInteractorsList}
\alias{getInteractorsList}
\title{Get a list of genes with their interactors}
\usage{
getInteractorsList(dataf, genes)
}
\arguments{
\item{dataf}{data frame obtained by the network}

\item{genes}{list of known regulators of ADAR}
}
\value{
list
}
\description{
This function is used to filter the data frame to consider only the
regulators of ADAR and, for each regulator, obtain a list of their
interactors and the corresponding scores.
}
\examples{

# Load the list of regulators
data(regulators_list)

# Load the example data frame
data(network_df)

# Get the gene symbols corresponding to the gene IDs in the data frame
gene_symbols <- getGeneSymbols(network_df, organism = "hsapiens")

# Get the converted data frame
converted_df <- getGeneSymbolDataframe(network_df, gene_symbols, 
organism = "hsapiens")

# Get the list of the regulators with their interactors
regulators <- getInteractorsList(converted_df, regulators_list)

}
