% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network}
\alias{network}
\title{Example Network}
\format{
Network
}
\source{
generated by ARACNe
}
\usage{
network
}
\description{
Data frame obtained from the network computed from the expression matrix.
}
\keyword{datasets}
