% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network_df}
\alias{network_df}
\title{Example Data Frame}
\format{
Data frame
\describe{
\item{from}{Character, gene ID}
\item{to}{Character, gene ID}
\item{weight}{Numeric, weight of the edge}
}
}
\source{
generated from the example expression matrix
}
\usage{
network_df
}
\description{
Data frame obtained from the network computed from the expression matrix.
}
\keyword{datasets}
