% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regulators_list}
\alias{regulators_list}
\title{Example Regulators List}
\format{
Data frame
\describe{
\item{gene}{Numeric, gene indicator}
\item{type}{Character, type of regulator}
\item{symbols}{Character, gene symbol}
}
}
\source{
generated for the example
}
\usage{
regulators_list
}
\description{
List of ADAR regulators.
}
\keyword{datasets}
