% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vsd_matrix}
\alias{vsd_matrix}
\title{Example Expression Matrix}
\format{
Matrix with rows representing genes and columns representing samples.
}
\source{
GEO - GSE54152
}
\usage{
vsd_matrix
}
\description{
Expression matrix with normalized values of a H. Sapiens dataset.
}
\keyword{datasets}
