% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getInteractorsList}
\alias{getInteractorsList}
\title{Get a list of genes with their interactors}
\usage{
getInteractorsList(dataf, reg_dataf, target_genes)
}
\arguments{
\item{dataf}{data frame obtained by the network}

\item{reg_dataf}{data frame with the information of the regulator genes}

\item{target_genes}{list of the genes of interest}
}
\value{
list
}
\description{
This function extracts from the data frame of regulators the list of the
gene symbols, and it adds the genes for which we want to compute the
activity with VIPER.
Then it filters the data frame obtained by the network to consider only the
regulators of the gene of interest and, for each regulator, it gets a list of
their interactors and the corresponding scores.
}
\examples{

# Load the list of regulators
data(regulators_list)

# Load the example data frame
data(network_df)

# Get the gene symbols corresponding to the gene IDs in the data frame
gene_symbols <- getGeneSymbols(network_df, organism = "hsapiens")

# Get the converted data frame
converted_df <- getGeneSymbolDataframe(network_df, gene_symbols,
organism = "hsapiens")

# Get the list of the regulators with their interactors
genes <- c("ADAR", "ADARB1")
interactors <- getInteractorsList(converted_df, regulators_list, genes)

}
