% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regulators_list}
\alias{regulators_list}
\title{Example Regulators List}
\format{
List
}
\source{
generated for the example
}
\usage{
regulators_list
}
\description{
List of ADAR regulators.
}
\keyword{datasets}
